refactor: styled component 컨벤션 적용
- .styled 파일 분리
- styled component 네이밍 컨벤션에 따라 네이밍 수정, https://github.com/Hi-Fi-Club/FE/wiki/Styled-Components-%3A-Naming-Convention
