+++ src/components/Header/Header.tsx
@@ -1,37 +1,17 @@
-import styled from 'styled-components';
 import PlainLink from '../PlainLink/PlainLink';
 import Logo from '../Logo/Logo';
-import RightMenu from './RightMenu';
+import RightMenu from '../RightMenu/RightMenu';
+import * as S from './Header.styled';
 
 function Header() {
   return (
-    <StyledHeader>
+    <S.HeaderBox>
       <PlainLink to="/">
         <Logo />
       </PlainLink>
       <RightMenu />
-    </StyledHeader>
+    </S.HeaderBox>
   );
 }
-
-const StyledHeader = styled.div`
-  padding: 0 10%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  height: 80px;
-  position: sticky;
-  top: 0px;
-
-  z-index: ${({ theme: { zPriorities } }) => zPriorities.overEverything};
-
-  color: ${({ theme: { colors } }) => colors.white};
-  background: ${({ theme: { colors } }) => colors.emerald};
-
-  ${({ theme: { media } }) => media.sm`
-    padding: 0 20px;
-  `};
-`;
 
 export default Header;
