+++ src/pages/CartPage/CartPage.tsx
@@ -1,10 +1,10 @@
-import styled from 'styled-components';
 import CartItem from '../../components/CartItem/CartItem';
-import CheckBox from '../../components/CheckBox/CheckBox';
+import Checkbox from '../../components/Checkbox/Checkbox';
 import Spinner from '../../components/Spinner/Spinner';
 import DivideLine from '../../components/DivideLine/DivideLine';
 import Button from '../../components/Button/Button';
 import useCart from './useCart';
+import * as S from './CartPage.styled';
 
 function CartPage() {
   const {
@@ -29,14 +29,14 @@ function CartPage() {
   }
 
   return (
-    <CartContainer>
-      <CartTitle>장바구니</CartTitle>
+    <S.PageBox>
+      <S.Title>장바구니</S.Title>
       <DivideLine />
-      <CartSectionsContainer>
-        <CartLeftSection>
-          <CartControlContainer>
-            <SelectAllContainer>
-              <CheckBox
+      <S.SectionsBox>
+        <S.LeftSection>
+          <S.ControlBox>
+            <S.SelectAllBox>
+              <Checkbox
                 id="select-all-checkbox"
                 name="select-all-checkbox"
                 checked={Object.values(checkedFlags).every(
@@ -45,12 +45,12 @@ function CartPage() {
                 onChange={handleCheckAll}
               />
               <label htmlFor="select-all-checkbox">선택해제</label>
-            </SelectAllContainer>
-            <RemoveSelectedButtonContainer onClick={removeAllCartItem}>
+            </S.SelectAllBox>
+            <S.RemoveSelectedButtonBox onClick={removeAllCartItem}>
               상품삭제
-            </RemoveSelectedButtonContainer>
-          </CartControlContainer>
-          <CartSubtitle>든든배송 상품 ({cart.length}개)</CartSubtitle>
+            </S.RemoveSelectedButtonBox>
+          </S.ControlBox>
+          <S.Subtitle>든든배송 상품 ({cart.length}개)</S.Subtitle>
           <DivideLine color="gray" />
           <ul>
             {cart.map(({ product, quantity }) => (
@@ -67,153 +67,29 @@ function CartPage() {
               </li>
             ))}
           </ul>
-        </CartLeftSection>
-        <CartRightSection>
-          <CartRightSectionTop>
-            <CartSubtitle>결제예상금액</CartSubtitle>
-          </CartRightSectionTop>
+        </S.LeftSection>
+        <S.RightSection>
+          <S.RightSectionTopBox>
+            <S.Subtitle>결제예상금액</S.Subtitle>
+          </S.RightSectionTopBox>
           <DivideLine thickness="thin" />
-          <TotalPriceContainer>
-            <HighlightText>결제예상금액</HighlightText>
-            <HighlightText>
+          <S.TotalPriceBox>
+            <S.HighlightedText>결제예상금액</S.HighlightedText>
+            <S.HighlightedText>
               {totalPrice.toLocaleString('ko-KR')}원
-            </HighlightText>
-          </TotalPriceContainer>
-          <OrderButtonContainer>
+            </S.HighlightedText>
+          </S.TotalPriceBox>
+          <S.OrderButtonBox>
             <Button>
               주문하기 (
               {Object.values(checkedFlags).filter((checked) => checked).length}
               개)
             </Button>
-          </OrderButtonContainer>
-        </CartRightSection>
-      </CartSectionsContainer>
-    </CartContainer>
+          </S.OrderButtonBox>
+        </S.RightSection>
+      </S.SectionsBox>
+    </S.PageBox>
   );
 }
-
-const CartContainer = styled.div`
-  width: 1200px;
-
-  ${({ theme: { media } }) => media.sm`
-    width: 100%;
-  `}
-
-  ${({ theme: { media } }) => media.md`
-    width: 100%;
-  `}
-`;
-
-const CartTitle = styled.h2`
-  font-size: 24px;
-  font-weight: 600;
-  text-align: center;
-`;
-
-const CartSectionsContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-
-  ${({ theme: { media } }) => media.sm`
-    flex-direction: column;
-  `}
-
-  ${({ theme: { media } }) => media.md`
-    flex-direction: column;
-  `}
-`;
-
-const CartLeftSection = styled.section`
-  width: 60%;
-  margin-top: 50px;
-
-  ${({ theme: { media } }) => media.sm`
-    width: 100%;
-  `}
-
-  ${({ theme: { media } }) => media.md`
-    width: 100%;
-  `}
-`;
-
-const CartControlContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: items-center;
-`;
-
-const SelectAllContainer = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 10px;
-`;
-
-const RemoveSelectedButtonContainer = styled.div`
-  padding: 12px 22px;
-  border: 1px solid #bbbbbb;
-`;
-
-const CartRightSection = styled.section`
-  width: 35%;
-  height: 260px;
-  margin-top: 50px;
-
-  border: 1px solid ${({ theme: { colors } }) => colors.gray};
-
-  ${({ theme: { media } }) => media.sm`
-    width: 100%;
-  `}
-  ${({ theme: { media } }) => media.md`
-    width: 100%;
-  `};
-`;
-
-const CartRightSectionTop = styled.div`
-  display: flex;
-  align-items: center;
-  padding: 16px 30px;
-`;
-
-const CartSubtitle = styled.h3`
-  display: flex;
-  align-items: center;
-  font-size: 20px;
-`;
-
-const TotalPriceContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  padding: 0 30px;
-  margin-top: 20px;
-`;
-
-const HighlightText = styled.span`
-  position: relative;
-  font-weight: 700;
-  display: inline-block;
-  text-align: center;
-  padding: 0 2px;
-  font-size: 20px;
-
-  ::after {
-    content: '';
-    display: block;
-    position: absolute;
-    left: 0;
-    bottom: 0;
-    width: 100%;
-    height: 8px;
-    background-color: #2ac1bc;
-    opacity: 0.5;
-    z-index: ${({ theme: { zPriorities } }) => zPriorities.behind};
-  }
-`;
-
-const OrderButtonContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 30px 30px 0 30px;
-`;
 
 export default CartPage;
