+++ src/styles/GlobalStyle.ts
@@ -17,6 +17,18 @@ const GlobalStyle = createGlobalStyle`
     cursor: pointer;
     padding: 0;
   }
+  
+  .visually-hidden {
+    position: absolute;
+    width: 1px;
+    height: 1px;
+    padding: 0;
+    margin: -1px;
+    overflow: hidden;
+    clip: rect(0, 0, 0, 0);
+    white-space: nowrap;
+    border-width: 0;
+  }
 `;
 
 export default GlobalStyle;
